requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
geographiclib==1.52
geopy==2.2.0
idna==3.3
paho-mqtt==1.6.1
psycopg2==2.9.3
requests==2.27
urllib3==1.26.9
python-dotenv==0.20.0